requirements.txt
azure-functions
azure-storage-blob
azure-identity